Move logger hook related code into separate file.
